pyproject.toml
[project]
name = "kid-mind"
version = "0.1.0"
description = "Platform for collecting and analysing EU PRIIPs Key Information Documents (KIDs) for ETFs"
authors = [
    {name = "oginskis"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "docling>=2.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
    "claude-agent-sdk>=0.1.46",
    "streamlit>=1.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.2.2",
    "chonkie[openai]>=1.5.6",
    "yfinance>=1.2.0",
    "requests>=2.32.5",
    "pydantic-ai>=1.61.0",
    "plotly>=5.18.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "opentelemetry-api>=1.20.0",
    "arize-phoenix-otel>=0.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kid_mind"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["kid_mind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
